+++ Cartfile.resolved
@@ -1,2 +1,2 @@
 github "s-aska/EventBox" "0.0.2"
-github "s-aska/KeyClip" "1.1.1"
+github "s-aska/KeyClip" "1.1.2"
